requirements.txt
asgiref==3.5.2
Django==3.2.16
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==1.4.6
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
typing-extensions==4.4.0
